Fix operating on multiple databases with `sqlite3Worker1Promiser`
The promiser implementation remembers the `dbId` returned by the worker on the
first `open` command. The `dbId` is then added to each following command if
none is given by the caller.

There are two flaws in the implementation without this fix:

1. An `open` command must not contain a `dbId`, since the `dbId` will be determined
by the worker as part of the command execution. However, on a second `open` command
the promiser will default to the `dbId` of the first `open` command. This change fixes
this by not setting a default `dbId` on `open` commands.
2. The promiser has two signatures, here given in TypeScript notation:

    // one parameter
    promiser(cmd: { type: string, dbId?: string, args: { /* command specific args */ })
    // two parameters
    promiser(type: string, args: { /* command specific args */ })

The issue lies in the second form. The `wrapper-worker.js` demo code suggests that the
`dbId` shall be supplied in the second argument. However, the `dbId` is not correctly read
from the second argument in the current implementation, so an empty `dbId` will be sent
to the worker, or the `dbId` from the first `open` command if one was executed. This
change fixes this by reading the `dbId` from the second argument object.